+++ src/s3ql/s3cache.py
@@ -9,6 +9,7 @@ from contextlib import contextmanager
 from s3ql import fs
 from s3ql.multi_lock import MultiLock
 from s3ql.ordered_dict import OrderedDict
+from s3ql.common import ExceptionStoringThread
 import errno
 import logging
 import os
@@ -69,8 +70,8 @@ class S3Cache(object):
       moved the file cursor.
  
     For this reason, all operations on s3 files are mediated by
-    an S3Cache object. Whenever the fs needs to write or read
-    from an s3key, it uses a context manager provided by the S3Cache
+    an S3Cache object. Whenever the file system needs to write or read
+    from an s3 object, it uses a context manager provided by the S3Cache
     object which returns a file handle to the s3 object. The S3Cache retrieves
     and stores objects on S3 as necessary. Moreover, it provides
     methods to delete and create s3 objects, once again taking care
@@ -102,14 +103,26 @@ class S3Cache(object):
     :timeout:     Maximum time to wait for changes in S3 to propagate
     :dbcm:        ConnectionManager instance, to manage access to the database
                   from different threads
+    :bgcommit:    Enable background commit mode
+    :commit_thread: Background committer thread
+    :shutdown:    Flag for committer thread to shut down
                   
     
     The `expect_mismatch` attribute is only for unit testing instrumentation
     and suppresses warnings if a mismatch between local and remote etag
     is encountered.                   
+    
+    Background Commit Mode
+    ----------------------
+    
+    In background commit mode, expire(), close() and flush() do not actually 
+    commit any data to S3. Instead, __init__() starts a background thread that
+    periodically performs cache expiration. This means that the cache can grow without
+    bounds and uncommitted changes may reside in the cache even after
+    the file system has been unmounted. 
     """
     
-    def __init__(self, bucket, cachedir, cachesize, dbcm, timeout=60):
+    def __init__(self, bucket, cachedir, cachesize, dbcm, timeout=60, bgcommit=False):
         log.debug('Initializing')
         self.keys = OrderedDict()
         self.cachedir = cachedir
@@ -121,20 +134,25 @@ class S3Cache(object):
         self.dbcm = dbcm
         self.timeout = timeout 
         self.expect_mismatch = False
-
+        self.bgcommit = bgcommit
+        self.shutdown = threading.Event()
+        self.shutdown.clear()
+        
+        # Reconstruct cache
+        self._recover_cache()
+             
+        if bgcommit:
+            # Start commit thread
+            self.commit_thread = ExceptionStoringThread(target=self._commit)
+            self.commit_thread.start()
+        
+            
     @contextmanager
     def get(self, inode, offset):
-        """Get filehandle for s3 object backing `inode` at `offset`
+        """Get file handle for s3 object backing `inode` at `offset`
         
-        This is a context manager function, so the intended usage is
-        
-        with s3cache.get(inode, offset, cur) as fh:
-            foo = fh.read(...)
-            
-        Note that offset has to be the starting offset. If caller is
-        going to write into the fh, he has to set `markdirty` so that
-        the changes are propagated back into S3.
-        
+        No checking is performed on the offset, the caller has to make
+        sure that it is at a block boundary.
         """
                  
         # Get s3 key    
@@ -197,13 +215,30 @@ class S3Cache(object):
             self.s3_lock.release(s3key)
             
             self.expire()
+      
+    def _recover_cache(self):
+        '''Read cache directory and register contents 
+        
+        Not synchronized. Must be called before other threads are running.
+        '''
+        
+        log.debug('Recovering cache files...')
+        for s3key in os.listdir(self.cachedir):
+            assert isinstance(s3key, unicode)
+            log.debug('Recovering %s', s3key)
+            el = CacheEntry(s3key, self.cachedir + s3key, "r+b")
+            el.dirty = True
+            self.keys[s3key] = el
             
+            el.seek(0, 2)
+            self.size += el.tell()
+              
+        log.debug('Cache recovery complete.')      
             
     def _download_object(self, s3key, etag):
         """Download s3 object from amazon into the cache. Return path.
         
         Not synchronized. 
-
         """
 
         log.debug('Attempting to download object %s from S3', s3key)
@@ -239,65 +274,104 @@ class S3Cache(object):
         log.debug('Object %s fetched successfully.', s3key)
         return cachepath      
         
-                  
     def expire(self):
-        """Performs cache expiry.
-
+        '''Perform cache expiry unless in background commit mode
+        
         If the cache is bigger than `self.cachesize`, the oldest
         entries are flushed until at least `self.blocksize`
-        bytes are available (or there are no objects left to flush).
-        """
+        bytes are available (or there is only one objects left to flush).
+        '''
+        
+        if self.bgcommit:
+            log.debug('Skipping cache expiration since bgcommit is enabled.')
+            return
+          
         log.debug('Expiring cache')
         
-        while self.size > self.maxsize:
-
-            with self.global_lock:
-                # If we pop the object before having locked it, another thread 
-                # may download it - overwriting the existing file!
-                try:
-                    el = self.keys.get_last()
-                except IndexError:
-                    # Nothing left to expire, this may happen despite the check
-                    # below due to race conditions
-                    break
-                
-                # We want to leave at least one object in the cache, otherwise
-                # we may end up storing & fetching the same object several times
-                # just for one read() or write() call
-                if len(self.keys) == 1:
-                    break
-                 
-                log.debug('Least recently used object is %s, obtaining object lock..', el.s3key)
-                self.s3_lock.acquire(el.s3key)  
-                
-            try:
-                try:
-                    del self.keys[el.s3key]
-                except KeyError:
-                    # Another thread already expired it, we need to try again
-                    continue
-
-                el.seek(0, 2)
-                size = el.tell()
-                el.close()
-                
-                if el.dirty:
-                    log.debug('Expiring dirty s3 object %s', el.s3key)
-                    etag = self.bucket.store_from_file(el.s3key, el.name)
-                    self.dbcm.execute("UPDATE s3_objects SET etag=?, size=? WHERE id=?",
-                                      (etag, size, el.s3key))
-                else:
-                    log.debug('Expiring unchanged s3 object %s', el.s3key)
-                    
-                # Update cachesize
-                self.size -= size
-                
-                os.unlink(el.name)
-
-            finally:
-                self.s3_lock.release(el.s3key)
-                
+        # We want to leave at least one object in the cache, otherwise
+        # we may end up storing & fetching the same object several times
+        # just for one read() or write() call
+        while self.size > self.maxsize and len(self.keys) > 1:
+            self._expire_entry()
+            
         log.debug("Expiration end")
+          
+    def _commit(self):
+        '''Run commit thread.
+        
+        In an infinite loop, we sleep 5 seconds and expire the cache
+        if necessary. 
+        '''           
+        
+        # Pylint ignore since we deliberately override the main logger
+        log = logging.getLogger("S3Cache.Committer") #pylint: disable-msg=W0621
+        log.debug('Started.')
+        while True:
+            while self.size > self.maxsize and len(self.keys) > 1:
+                if self.shutdown.is_set():
+                    log.debug('Received shutdown signal, exiting.')
+                    return
+                
+                log.debug('Expiring entry..')
+                self._expire_entry()
+                
+            # Wait for 5 seconds or until shutdown is set
+            log.debug('Nothing more to expire. Sleeping...')
+            self.shutdown.wait(5)
+            
+            if self.shutdown.is_set():
+                log.debug('Received shutdown signal, exiting.')
+                return
+            
+    
+            
+    def _expire_entry(self):
+        """Remove the oldest entry from the cache.
+        """
+         
+        log.debug('Trying to expire oldest cache entry..') 
+        
+        with self.global_lock:
+            # If we pop the object before having locked it, another thread 
+            # may download it - overwriting the existing file!
+            try:
+                el = self.keys.get_last()
+            except IndexError:
+                log.debug('No objects in cache. Returning.')
+                return
+            
+            log.debug('Least recently used object is %s, obtaining object lock..', el.s3key)
+            self.s3_lock.acquire(el.s3key)  
+            
+        try:
+            try:
+                del self.keys[el.s3key]
+            except KeyError:
+                # Another thread already expired it, we need to try again
+                log.debug('Object has already been expired in another thread, returning.')
+                return
+
+            el.seek(0, 2)
+            size = el.tell()
+            el.close()
+            
+            if el.dirty:
+                log.debug('Commiting dirty s3 object %s...', el.s3key)
+                etag = self.bucket.store_from_file(el.s3key, el.name)
+                self.dbcm.execute("UPDATE s3_objects SET etag=?, size=? WHERE id=?",
+                                  (etag, size, el.s3key))
+   
+            log.debug('Removing s3 object %s from cache..', el.s3key)
+                
+            # Update cachesize
+            self.size -= size
+            
+            os.unlink(el.name)
+
+        finally:
+            self.s3_lock.release(el.s3key)
+                
+        
 
     def remove(self, inode, offset=0):
         """Unlinks all s3 objects from the given inode.
@@ -358,10 +432,15 @@ class S3Cache(object):
         
         
     def flush(self, inode):
-        """Uploads all dirty data from `inode` to S3
+        """Uploads all dirty data from `inode` to S3 unless in background commit mode
         
         No locking required. 
         """
+        
+        if self.bgcommit:
+            log.debug('Skipping flush for inode %d since in bgcommit mode', inode)
+            return
+        
         # Determine s3 objects from this inode
         log.debug('Flushing objects for inode %i', inode) 
         with self.dbcm() as conn:
@@ -385,26 +464,47 @@ class S3Cache(object):
                             (etag, el.tell(), el.s3key))
                 el.dirty = False
                 
+        log.debug('Flushing for inode %d completed.', inode)
 
     def close(self):
         """Uploads all dirty data and cleans the cache.
              
+        If background commit is activated, the committing thread is 
+        stopped and all committed data is cleaned from the cache. 
+        Uncommitted data remains in the cache.
         """       
 
         log.debug('Closing S3Cache') 
+        
+        if self.shutdown.is_set():
+            raise RuntimeError("close was called more than once!")
+        
+        self.shutdown.set()
+        if self.bgcommit:
+            log.debug('Waiting for background thread to exit..')
+            self.commit_thread.join_and_raise()
+            log.debug('Background thread has exited.')
+            
         with self.global_lock:
             while len(self.keys):
                 el = self.keys.pop_last()
                 
-                if el.dirty:
+                
+                if not el.dirty:
+                    el.close()
+                    os.unlink(el.name)
+                elif self.bgcommit:
+                    el.close()
+                else:
                     el.flush()    
                     el.seek(0, 2)
                     etag = self.bucket.store_from_file(el.s3key, el.name)
                     self.dbcm.execute("UPDATE s3_objects SET etag=?, size=? WHERE id=?",
                                 (etag, el.tell(), el.s3key))
 
-                el.close()
-                os.unlink(el.name)
+                    el.close()
+                    os.unlink(el.name)
+
                 
     def __del__(self):
         if self.keys:
